+++ data-raw/create_constants.R
@@ -420,11 +420,15 @@ eiou_flows <- list(bkb_peat_briquette_plants = "BKB/peat briquette plants",
                    gas_to_liquids_plants = "Gas-to-liquids (GTL) plants",
                    gasification_plants = "Gasification plants for biogases",
                    liquefaction_regasification_plants = "Liquefaction (LNG) / regasification plants",
+                   natural_gas_extraction = "Natural gas extraction",
                    non_specified_eiou = "Non-specified (energy)",
                    nuclear_industry = "Nuclear industry",
                    oil_and_gas_extraction = "Oil and gas extraction",
                    oil_refineries = "Oil refineries",
                    own_use_elect_chp_heat_plants = "Own use in electricity, CHP and heat plants",
+                   # In several places, we route "Own use in electricity, CHP and heat plants" to "Main activity producer electricity plants",
+                   # so need to include those plants here.
+                   main_activity_producer_electricity_plants = "Main activity producer electricity plants",
                    patent_fuel_plants = "Patent fuel plants",
                    pumped_storage_plants = "Pumped storage plants")
 usethis::use_data(eiou_flows, overwrite = TRUE)
@@ -550,6 +554,8 @@ usethis::use_data(aggregation_flows, overwrite = TRUE)
 # Default names for columns in aggregate data frames
 #
 aggregate_cols <- list(aggregate_primary = "EX.p",
+                       aggregate_final = "EX.f",
+                       aggregate_useful = "EX.u",
                        net_aggregate_demand = "EX.fd_net",
                        gross_aggregate_demand = "EX.fd_gross")
 usethis::use_data(aggregate_cols, overwrite = TRUE)